remove dependency on aws sdk v1
It was only being used in test cases for the SES helper
library. That usage has been converted to v2.
